requirements.txt
fastapi
uvicorn
twilio
openai
python-dotenv
python-multipart
